feat: 22.4 learning exercises update
